feat(create-gateway): 创建网关页面改版 & ai 网关对接
