+++ src/dashboard-front/src/locales/cn.json
@@ -444,7 +444,7 @@
   "网关类型": "网关类型",
   "查看指引 >": "查看指引 >",
   "普通网关发布流程": "普通网关发布流程",
-  "编程网关发布流程": "编程网关发布流程",
+  "可编程网关发布流程": "可编程网关发布流程",
   "初始化环境： prod ，并可新增环境": "初始化环境： prod ，并可新增环境",
   "配置后端服务": "配置后端服务",
   "每个环境都可以单独配置后端 API 的域名": "每个环境都可以单独配置后端 API 的域名",
@@ -473,6 +473,12 @@
   "通过编码的方式配置和管理网关的功能，支持接口组合、协议转换和接口编排等功能": "通过编码的方式配置和管理网关的功能，支持接口组合、协议转换和接口编排等功能",
   "自动创建开源仓库，将模板代码初始化到仓库中，并将创建者设定为仓库管理员": "自动创建开源仓库，将模板代码初始化到仓库中，并将创建者设定为仓库管理员",
   "组件配置有更新，新增组件或更新组件请求方法、请求路径、权限级别、用户认证，需同步到网关才能生效": "组件配置有更新，新增组件或更新组件请求方法、请求路径、权限级别、用户认证，需同步到网关才能生效",
+  "AI 网关发布流程": "AI 网关发布流程",
+  "配置模型服务": "配置模型服务",
+  "按环境配置 OpenAI 兼容的 endpoint、auth、model 等，并做连通测试": "按环境配置 OpenAI 兼容的 endpoint、auth、model 等，并做连通测试",
+  "新建「模型代理 API」引用模型服务，或新建普通资源": "新建「模型代理 API」引用模型服务，或新建普通资源",
+  "资源列表快照可以用于对比差异": "资源列表快照可以用于对比差异",
+  "发布后下发 APISIX，即可通过网关受控调用模型": "发布后下发 APISIX，即可通过网关受控调用模型",
   "新建组件": "新建组件",
   "模型服务": "模型服务",
   "后端/模型服务": "后端/模型服务",
