+++ src/dashboard-front/src/locales/en.json
@@ -444,7 +444,7 @@
   "网关类型": "Gateway type",
   "查看指引 >": "view guide >",
   "普通网关发布流程": "Common Gateway Release Process",
-  "编程网关发布流程": "Programming gateway Release Process",
+  "可编程网关发布流程": "Programming gateway Release Process",
   "初始化环境： prod ，并可新增环境": "Initialize environment: prod and add environment",
   "配置后端服务": "Configure back-end services",
   "每个环境都可以单独配置后端 API 的域名": "Each environment can be individually configured with the domain name of the back-end API",
@@ -473,6 +473,12 @@
   "通过编码的方式配置和管理网关的功能，支持接口组合、协议转换和接口编排等功能": "Configure and manage the functions of the gateway by means of coding, supporting interface composition, protocol conversion, interface orchestration and other functions",
   "自动创建开源仓库，将模板代码初始化到仓库中，并将创建者设定为仓库管理员": "Automatically create an open source repository, initialize the template code into the repository, and set the creator as the repository administrator",
   "组件配置有更新，新增组件或更新组件请求方法、请求路径、权限级别、用户认证，需同步到网关才能生效": "The components have been updated, may add components, or update request method, path, permission level, user verified, need to synchronize to the gateway to take effect",
+  "AI 网关发布流程": "AI Gateway Release Process",
+  "配置模型服务": "Configure model service",
+  "按环境配置 OpenAI 兼容的 endpoint、auth、model 等，并做连通测试": "Configure OpenAI-compatible endpoint, auth, model, and perform connectivity testing by environment",
+  "新建「模型代理 API」引用模型服务，或新建普通资源": "Create a 「Model Proxy API」 to reference the model service, or create a normal resource",
+  "资源列表快照可以用于对比差异": "Resource list snapshots can be used to compare differences",
+  "发布后下发 APISIX，即可通过网关受控调用模型": "Once published, APISIX will be issued, and the model can be called through the gateway in a controlled manner",
   "新建组件": "Create Component",
   "模型服务": "Model Service",
   "后端/模型服务": "Backend/Model Service",
