+++ src/dashboard-front/src/router/index.ts
@@ -60,6 +60,16 @@ const routes: RouteRecordRaw[] = [
     component: () => import('@/views/home/Index.vue'),
   },
   {
+    path: '/create-gateway',
+    name: 'CreateGateway',
+    component: () => import('@/views/home/CreateGateway.vue'),
+  },
+  {
+    path: '/gateway-guide/:id/:name',
+    name: 'GatewayGuide',
+    component: () => import('@/views/home/GatewayGuide.vue'),
+  },
+  {
     path: '/:id',
     name: 'Resources',
     component: () => import('@/layout/my-gateway/Index.vue'),
